# openapi.yaml
openapi: 3.0.0
...
info:
  title: DBS Library API
  version: '1.0'
  description: 'This API provides endpoints for managing users, RFID cards, books, book copies, and book rentals with reservations.'
  contact:
    name: Jakub Dubec
    email: jakub.dubec@stuba.sk
servers:
  - url: 'http://localhost:8000'
...
paths:
  '/users/{userId}':
    parameters:
      - schema:
          type: string
          format: uuid
        name: userId
        in: path
        required: true
        description: Id of an existing user.
    get:
      summary: Get User Info by User ID
      tags:
        - Users
      responses:
        '200':
          description: User Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              examples:
                Example 1:
                  value:
                    id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                    name: string
                    surname: string
                    email: user@example.com
                    birth_date: '2019-08-24'
                    personal_identificator: string
                    reservations:
                      - id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                        user_id: a169451c-8525-4352-b8ca-070dd449a1a5
                        publication_id: c3452ed9-df18-461f-a6fe-5036c6b93bf2
                        created_at: '2019-08-24T14:15:22Z'
                    rentals:
                      - id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                        user_id: a169451c-8525-4352-b8ca-070dd449a1a5
                        publication_instance_id: d53415a1-1c04-429f-a852-288aba36578a
                        duration: 0
                        start_date: '2019-08-24T14:15:22Z'
                        end_date: '2019-08-24T14:15:22Z'
                        status: active
                    created_at: '2019-08-24T14:15:22Z'
                    updated_at: '2019-08-24T14:15:22Z'
        '404':
          description: User Not Found
      operationId: get-users-userId
      description: Retrieve the information of the user with the matching user ID.
      parameters: []
    patch:
      summary: Update User Information
      operationId: patch-users-userId
      responses:
        '200':
          description: User Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              examples: {}
        '400':
          description: Bad Request
        '404':
          description: User Not Found
        '409':
          description: Email Already Taken
      description: Update the information of an existing user.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              x-examples:
                Example 1:
                  name: string
                  surname: string
                  email: user@example.com
                  birth_date: '2019-08-24'
                  personal_identificator: string
              properties:
                name:
                  type: string
                surname:
                  type: string
                email:
                  type: string
                  format: email
                birth_date:
                  type: string
                  format: date
                personal_identificator:
                  type: string
            examples:
              Example 1:
                value:
                  name: string
                  surname: string
                  email: user@example.com
                  birth_date: '2019-08-24'
                  personal_identificator: string
        description: Patch user properties to update.
      tags:
        - Users
...
components:
  schemas:
    User:
      title: User
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        surname:
          type: string
        email:
          type: string
          format: email
        birth_date:
          type: string
          format: date
        personal_identificator:
          type: string
        reservations:
          type: array
          items:
            $ref: '#/components/schemas/Reservation'
        rentals:
          type: array
          items:
            $ref: '#/components/schemas/Rental'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - surname
        - email
        - birth_date
        - personal_identificator
        - created_at
        - updated_at
      description: 'This JSON object represents the complete detail of a user in the library system. It includes the user ID, name, surname, email, birth date, and personal identifier. It also contains a list of the active reservations and rentals, ordered by the date they were created.'
...
    Rental:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        publication_instance_id:
          type: string
          format: uuid
        duration:
          type: integer
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - active
            - returned
            - overdue
      description: |
        - `id`: A string representing the unique UUID of the newly created rental.
        - `user_id`: A string representing the UUID of the user who created the rental.
        - `publication_instance_id`: A string representing the UUID of the publication instanceassigned to the rental.
        - `duration`: An integer representing the duration of the rental in days.
        - `start_date`: A string representing the date and time when the rental begins, in ISO 8601 format (e.g. "2023-04-11T12:00:00Z").
        - `end_date`: A string representing the date and time when the rental is due to end, inISO 8601 format (e.g. "2023-04-18T12:00:00Z").
        - `status`: A string representing the current status of the rental. Valid values include "active", "overdue", and "returned".

        The `id`, `user_id`, `publication_instance_id`, and `duration` properties are as specified in the original request. The `start_date` property is automatically set to the current date and time when the rental is created, and the `end_date` property is calculated based on the duration and the current date and time.

        The `status` property indicates the current state of the rental. The rental is initially "active" when created, and may transition to "overdue" if it is not returned by the due date, or "returned" if it is successfully returned.
    Reservation:
      type: object
      x-examples:
        Example 1:
          id: 8f02a2c2-6477-4f63-9658-14d39e1f7c1e
          user_id: 3d2737f5-b671-43a7-b8fd-4039a5ad2e6b
          publication_id: 0d68a4b5-5a63-4a5d-9575-5ef3a3aa2c5b
          created_at: '2023-04-15T09:30:00Z'
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        publication_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
      description: |
        - `id`: A string representing the unique UUID of the new reservation.
        - `user_id`: A string representing the UUID of the user who created the reservation.
        - `publication_id`: A string representing the UUID of the publication for which the reservation was created.
        - `created_at`: A string representing the timestamp when the reservation was created.